+++ R/selangle.R
@@ -1,15 +1,15 @@
 selangle <- function(
-    X, Y = NULL, ncomp = NULL, algo = NULL, 
+    X, Y = NULL, ncomp = NULL, algo = NULL,
     B = 50, seed = NULL,
-    type = c("R", "P"),
+    angle = c("maxsub", "hot"),
     plot = TRUE, 
     xlab = "Nb. components", ylab = NULL,
     print = TRUE, 
     ...
     ) {
     
-    type <- match.arg(type)
-    
+    angle <- match.arg(angle)   
+  
     X <- .matrix(X)
     zdim <- dim(X)
     n <- zdim[1]
@@ -30,8 +30,6 @@ selangle <- function(
         if(is.null(algo))
             algo <- pls_kernel
         fm <- algo(X, Y, ncomp = ncomp, ...)
-        if(type == "R")
-            fm$P <- fm$R
         }
     
     set.seed(seed = seed)
@@ -48,14 +46,11 @@ selangle <- function(
             Y <- .matrix(Y, row = FALSE, prefix.colnam = "y")
             zfm <- algo(X[s, ], Y[s, ], ncomp = ncomp)
             }
-        
-        if(type == "R")
-            zfm$P <- zfm$R
 
         zq <- numeric()
         for(a in seq_len(ncomp))
             zq[a] <- .corvec(fm$P[, seq_len(a), drop = FALSE],
-                             zfm$P[, seq_len(a), drop = FALSE], type = "hot")
+                             zfm$P[, seq_len(a), drop = FALSE], type = angle)
 
         Q[i, ] <- zq
         
@@ -70,8 +65,9 @@ selangle <- function(
     
     ## Output in radiants: right angle = pi/2)
     ## Radiants are then standardized to [0, 1] by dividing by pi/2 
-    r <- acos(q) / (pi / 2)
-    r[r == 0] <- 1e-4
+    q <- acos(q) / (pi / 2)
+    
+    q[q == 0] <- 1e-4
     
     if(plot) {
       
@@ -81,11 +77,11 @@ selangle <- function(
         oldpar <- par(mfrow = c(1, 1))
         par(mfrow = c(1, 2))
         
-        .plot_scree(r, ylim = c(.9 * min(r), 1.01),
+        .plot_scree(q, ylim = c(.9 * min(q), 1.01),
                     xlab = xlab, ylab = ylab)
         abline(h = 1, col = "grey")
         
-        .plot_scree(log(r), ylim = c(1.1 * min(log(r)), .02), 
+        .plot_scree(log(q), ylim = c(1.1 * min(log(q)), .02), 
                     xlab = xlab, ylab = ylab, main = "log-scale")
         abline(h = 0, col = "grey")
         
@@ -93,7 +89,7 @@ selangle <- function(
 
         }
     
-    list(r = r)
+    list(q = q)
     
     }